Only reverse geocode the location after the user drags the map
Previously, if a user searched for a location in the geocode box, the location clicked would be set in the idea form, but then the map would reverse geocode the center to an exact address because the reverse geocoding was tied to the mapmoveend event. Since the map moves after a geocode option is selected, the geocoded location would get overridden by the reverse geocoded address. When the user did something like select a neighborhood, this is undesireable behavior